moved settings to another tab; updated README
